cleaned up subpackages and imports
