+++ modules/__init__.py
@@ -1,3 +1,15 @@
+__all__ = ["base",
+    "http",
+    "snmp",
+    "munin",
+    "ping",
+    "cmd",
+    "printer",
+    "file",
+    "conversion",
+    "websocket",
+]
+
 from .base import *
 from .http import *
 from .snmp import *
