+++ modules/file.py
@@ -1,6 +1,6 @@
 import logging
 
-from .base import TimedSource
+from .base.sources import TimedSource
