+++ modules/printer.py
@@ -1,4 +1,4 @@
-from .base import PubSubSource
+from .base.sources import PubSubSource
