+++ modules/snmp.py
@@ -1,7 +1,7 @@
 import logging
 import subprocess
 
-from .base import TimedSource
+from .base.sources import TimedSource
